+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Wang 2019 ceftiofur ([doi:10.3389/fvets.2019.00363](https://doi.org/10.3389/fvets.2019.00363)) -- healthy beagle dogs.
+- Add Bhagunde 2019 imipenem and relebactam ([doi:10.1002/psp4.12462](https://doi.org/10.1002/psp4.12462)) -- healthy adults and adults with complicated intra-abdominal infection, complicated urinary tract infection, or hospital-acquired/ventilator-associated bacterial pneumonia.
 
 - Add Palmer 2025 moxifloxacin ([doi:10.1002/bcp.70005](https://doi.org/10.1002/bcp.70005)) -- children with rifampicin-resistant tuberculosis.